requirements.txt
google-generativeai
google-adk
python-dotenv
requests
fastapi
uvicorn[standard]
pydantic
tiktoken
